refactor: remove unused hero variation components and update localization files
- Deleted Hero3D, HeroBento, HeroBrutalist, HeroGamer, HeroGlass, HeroMagazine, HeroOrganic, HeroScroll, HeroTerminal, and HeroZen components.
- Updated localization files (ca.json, en.json, es.json) to include new keys for "neoTitle", "neoSubtitle", "neoCta", and "neoBadge".
