+++ src/i18n/locales/ca.json
@@ -15,7 +15,11 @@
         "subtitle": "Desenvolupador Frontend especialitzat en React amb experiència creant interfícies modernes i funcionals",
         "cta": "Veure el meu treball",
         "downloadCV": "Descarregar CV",
-        "viewProjects": "Veure Projectes"
+        "viewProjects": "Veure Projectes",
+        "neoTitle": "WEBCODE.ES",
+        "neoSubtitle": "CREANT CAOS I ORDRE DIGITAL",
+        "neoCta": "VEURE PROJECTES",
+        "neoBadge": "WEB CODE"
     },
     "experience": {
         "title": "Experiència Professional",
